+++ R/cnd-message.R
@@ -136,6 +136,9 @@ cnd_footer.default <- function(cnd, ...) {
 #' # All bullets
 #' writeLines(format_bullets(c("foo", "bar")))
 #'
+#' # This is equivalent to
+#' writeLines(format_bullets(set_names(c("foo", "bar"), "*")))
+#'
 #' # Supply named elements to format info, cross, and tick bullets
 #' writeLines(format_bullets(c(i = "foo", x = "bar", v = "baz", "*" = "quux")))
 #'